# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/feeds/nearby": {
      "get": {
        "operationId": "FeedsController_nearby_v1",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "example": 33.7531,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "schema": {
              "example": -84.3857,
              "type": "number"
            }
          },
          {
            "name": "radiusMiles",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 5,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque pagination cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Published deals near a point, sorted by distance (cursor paginated)",
        "tags": [
          "feeds"
        ]
      }
    },
    "/v1/feeds/recommended": {
      "get": {
        "operationId": "FeedsController_recommended_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Personalized, explainable recommendations (with reasons)",
        "tags": [
          "feeds"
        ]
      }
    },
    "/v1/feeds/trending": {
      "get": {
        "operationId": "FeedsController_trending_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Trending deals by recent popularity",
        "tags": [
          "feeds"
        ]
      }
    },
    "/v1/deals/{id}/swipes": {
      "post": {
        "operationId": "ActionsController_swipe_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSwipeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Record a swipe (right saves). Supports Idempotency-Key.",
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/deals/{id}/swipes/latest": {
      "delete": {
        "operationId": "ActionsController_undo_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Undo the latest swipe for this deal",
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/deals/{id}/save": {
      "post": {
        "operationId": "ActionsController_save_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "actions"
        ]
      },
      "delete": {
        "operationId": "ActionsController_unsave_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/deals/{id}/watch": {
      "post": {
        "operationId": "ActionsController_watch_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "actions"
        ]
      },
      "delete": {
        "operationId": "ActionsController_unwatch_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/deals/{id}/views": {
      "post": {
        "operationId": "ActionsController_view_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/deals/{id}/clicks": {
      "post": {
        "operationId": "ActionsController_click_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/deals/{id}/shares": {
      "post": {
        "operationId": "ActionsController_share_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/deals/{id}/redemptions": {
      "post": {
        "operationId": "ActionsController_redeem_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Mark used — counts realized savings once per user+deal",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/v1/me/entitlements": {
      "get": {
        "operationId": "SubscriptionsController_entitlements_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Server-computed entitlements (Dealy+)",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/v1/subscriptions/apple/sync": {
      "post": {
        "operationId": "SubscriptionsController_sync_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleSyncDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Verify a StoreKit transaction and refresh entitlements",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/v1/webhooks/apple": {
      "post": {
        "operationId": "SubscriptionsController_webhook_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "App Store Server Notifications v2 webhook",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/v1/admin/users/{id}/roles": {
      "post": {
        "operationId": "AdminController_grantRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Grant a role to a user (audited)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/deals/{id}/publish": {
      "post": {
        "operationId": "AdminController_publish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/deals/{id}/unpublish": {
      "post": {
        "operationId": "AdminController_unpublish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/deals/{id}/expire": {
      "post": {
        "operationId": "AdminController_expire_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ingestion/failures": {
      "get": {
        "operationId": "AdminController_failures_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Recent ingestion failures",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/audit-logs": {
      "get": {
        "operationId": "AdminController_auditLogs_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Recent audit log entries",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "Dealy API",
    "description": "Dealy production API — swipe-first, location-aware deals.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "CreateSwipeDto": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "enum": [
              "left",
              "right",
              "up"
            ],
            "description": "right also saves the deal"
          }
        },
        "required": [
          "direction"
        ]
      },
...
      "AppleSyncDto": {
        "type": "object",
        "properties": {}
      },
      "AppleWebhookDto": {
        "type": "object",
        "properties": {}
      },
      "GrantRoleDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}